requirements.txt
Django>=5.0
djangorestframework
numpy
pandas
scikit-learn
matplotlib
openpyxl
gunicorn
whitenoise
pip freeze > requirements.txt